packages/omega-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omega-orchestrator"
version = "0.1.0"
description = "Safety-first orchestration runtime implementing leak-first gates, conservation budgets, ethics thresholds, and temporal guards."
authors = [{name = "Steven Lanier"}, {name="SEAL Division"}]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
dependencies = ["pyyaml>=6.0", "pydantic>=2.6"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[project.scripts]
omega = "omega_orchestrator.cli:main"
